Remove extraneous `requestFrom(int, int, int)` calls
The lines changed cause problems on the STM32 platform, where they cause an error due to the unavailability of the `requestFrom(int, int, int)` signature. 

At any rate, the way this library uses the function call, the problem third parameter is extraneous, as the Arduino implementation of `requestFrom(int, int)` behaves exactly the same way as it would with `requestFrom(int, int, int)` with the third parameter being `true`.